Refactor RandomChordPractice component for better maintainability
- Split large components into focused, reusable pieces:
  * Extract ChordDisplay component (40 lines)
  * Extract MetronomeControls component (51 lines)
  * Extract BackNavigation component (23 lines)
  * Extract RandomModeSettings component (156 lines)
  * Extract DiatonicModeSettings component (88 lines)

- Create custom hooks for business logic separation:
  * useChordGeneration hook - chord generation logic
  * useMetronome hook - metronome state and timing
  * usePracticeSettings hook - localStorage state management

- Eliminate code duplication:
  * Create shared navigation configuration
  * Update navigation components to use shared config

- Add performance optimizations:
  * React.memo for component memoization
  * useCallback for event handlers
  * useMemo for expensive calculations
  * Optimized useEffect dependencies

- Reduce component complexity:
  * RandomChordPractice: 252→94 lines
  * Settings: 350→120 lines

All functionality preserved with improved code organization.
