+++ config/navigation.ts
@@ -1,0 +1,14 @@
+export const menuItems = [
+  {
+    title: 'Exercises',
+    href: '/exercises',
+  },
+  {
+    title: 'About',
+    href: '/about',
+  },
+  {
+    title: 'Contact',
+    href: '/contact',
+  },
+]
